requirements.txt
fastapi
uvicorn
praw
google-generativeai
python-dotenv
textblob
pydantic
